requirements.txt
openai
fastapi
aiohttp
python-dotenv
uvicorn
requests